+++ src/Control/Carrier/Cut/Church.hs
@@ -86,7 +86,7 @@ instance MonadTrans CutC where
   lift m = CutC (\ cons nil _ -> m >>= flip cons nil)
   {-# INLINE lift #-}
 
-instance (Algebra sig m, CanHandle sig (CutC m)) => Algebra (Cut :+: NonDet :+: sig) (CutC m) where
+instance (Algebra sig m, CanThread sig (CutC m)) => Algebra (Cut :+: NonDet :+: sig) (CutC m) where
   alg (L Cutfail)    = CutC $ \ _    _   fail -> fail
   alg (L (Call m k)) = CutC $ \ cons nil fail -> runCut (\ a as -> runCut cons as fail (k a)) nil nil m
   alg (R (L (L Empty)))      = empty