+++ src/Control/Carrier/Fresh/Strict.hs
@@ -52,7 +52,7 @@ evalFresh n (FreshC m) = evalState n m
 newtype FreshC m a = FreshC (StateC Int m a)
   deriving (Alternative, Applicative, Functor, Monad, Fail.MonadFail, MonadFix, MonadIO, MonadPlus, MonadTrans)
 
-instance (Algebra sig m, CanHandle sig ((,) Int)) => Algebra (Fresh :+: sig) (FreshC m) where
+instance (Algebra sig m, CanThread sig ((,) Int)) => Algebra (Fresh :+: sig) (FreshC m) where
   alg (L (Fresh k)) = FreshC (get <* modify (+ (1 :: Int))) >>= k
   alg (R other)     = FreshC (handleCoercible other)
   {-# INLINE alg #-}