+++ test/Effect.hs
@@ -30,11 +30,11 @@ data All m k = forall a . All (m a) ([a] -> m k)
 
 deriving instance Functor m => Functor (All m)
 
--- | Note that since the continuation takes @[a]@, our 'Effect' instance will have to produce a continuation taking @[f a]@ given one taking @[a]@, for which purpose we use 'sequenceA'. That necessitates an 'Applicative' instance for @f@, so we specialize 'CanHandle' for this effect to strengthen the constraint.
+-- | Note that since the continuation takes @[a]@, our 'Effect' instance will have to produce a continuation taking @[f a]@ given one taking @[a]@, for which purpose we use 'sequenceA'. That necessitates an 'Applicative' instance for @f@, so we specialize 'CanThread' for this effect to strengthen the constraint.
 --
 -- Intuitively, we can think of this as saying that if there are 'Control.Effect.State.State' effects present in @m@, collecting all of their results—joining the branches—requires us to merge the states monoidally; if there are 'Control.Effect.Error.Error' effects, it requires that they’ve all succeeded.
 instance Effect All where
-  type CanHandle All f = Applicative f
+  type CanThread All f = Applicative f
   handle ctx hdl (All m k) = All (hdl (m <$ ctx)) (hdl . fmap k . sequenceA)
 
 all :: Has All sig m => m a -> m [a]
@@ -60,10 +60,10 @@ data Thread m k
   | Yield (m k)
   deriving (Functor)
 
--- | This demonstrates that we can use '~' to define 'CanHandle', enforcing the absence of state or control that would get dropped.
+-- | This demonstrates that we can use '~' to define 'CanThread', enforcing the absence of state or control that would get dropped.
 --
 -- We don’t go as far as to define a carrier because cooperative multithreading is a bit more of an example than a test, and I don’t care to run 'Control.Concurrent.forkIO' in the unit tests. At any rate, the existence of this instance is test enough for what we’re trying to do.
 instance Effect Thread where
-  type CanHandle Thread f = f ~ Identity
+  type CanThread Thread f = f ~ Identity
   handle ctx hdl (Fork m k) = Fork (runIdentity <$> hdl (m <$ ctx)) (hdl (k <$ ctx))
   handle ctx hdl (Yield  k) = Yield                                 (hdl (k <$ ctx))